+++ src/Exception/DigestGenerationException.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Becklyn\CertKeyChecker\Exception;
+
+final class DigestGenerationException extends \RuntimeException implements CertKeyCheckerException
+{
+}