+++ src/Exception/MultipleFilesOfTypeException.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Becklyn\CertKeyChecker\Exception;
+
+final class MultipleFilesOfTypeException extends \RuntimeException implements CertKeyCheckerException
+{
+}