+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 =====
 
 *   (improvement) Bump dependencies to PHP 7.4 and Symfony 5.
+*   (bug) Fix handling of invalid files.
 
 
 1.0.0